Changes to energy types

+++ Assets/Scripts/Entity/Goblin.cs
@@ -6,7 +6,7 @@ using UnityEngine.AI;
 public class Goblin : MonoBehaviour
 {
     private NavMeshAgent agent;
-    private int health = 5; //TODO: Config!
+    private float health = 5; //TODO: Config!
     private float agentSpeed; //TODO: Config!
     private float freezed = 0f;
     private float freezeStarted = 0f;
